accordion refactor
