+++ src/components/common/accordion/types/index.ts
@@ -1,0 +1,17 @@
+type HeaderProps = {
+  children: React.ReactNode;
+  className?: string;
+  id?: string;
+};
+
+type BodyProps = {
+  children: React.ReactNode;
+  className?: string;
+};
+
+type AccordionProps = {
+  children: React.ReactNode;
+  className?: string;
+};
+
+export { AccordionProps, BodyProps, HeaderProps };
